refactor: FileModalScreen
