+++ src/view/pages/FileModalScreen/index.stories.tsx
@@ -1,16 +1,21 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react-native'
-import { FileModalScreenView } from '.'
+import { MainView } from './MainView'
 import { File } from '../../../models'
 
 const imageFile1 = new File(1, 'iona_v2.png', 'image', [
   {
+    id: 0,
+    score: 0,
+    extension: 'png',
     type: 'thumbnail',
     size: 0,
     mime: 'image/png',
     url: 'https://static.tosukeapps.tk/iona_v2_x128.png',
   },
   {
+    id: 1,
+    score: 0,
     extension: 'png',
     type: 'image',
     size: 0,
@@ -21,12 +26,17 @@ const imageFile1 = new File(1, 'iona_v2.png', 'image', [
 
 const imageFile2 = new File(2, 'iona_v4.png', 'image', [
   {
+    id: 2,
+    score: 0,
+    extension: 'png',
     type: 'thumbnail',
     size: 0,
     mime: 'image/png',
     url: 'https://static.tosukeapps.tk/iona_v4_x128.png',
   },
   {
+    id: 3,
+    score: 0,
     extension: 'png',
     type: 'image',
     size: 0,
@@ -37,12 +47,18 @@ const imageFile2 = new File(2, 'iona_v4.png', 'image', [
 
 const videoFile = new File(3, 'stack.mp4', 'video', [
   {
+    id: 4,
+    score: 0,
+    extension: 'png',
     type: 'thumbnail',
     size: 0,
     mime: 'image/png',
     url: 'https://i.imgur.com/3AHexTe.png',
   },
   {
+    id: 5,
+    score: 0,
+    extension: 'mp4',
     type: 'video',
     size: 0,
     mime: 'video/mp4',
@@ -51,5 +67,5 @@ const videoFile = new File(3, 'stack.mp4', 'video', [
 ])
 
 storiesOf('FileModalScreenView', module)
-  .add('image', () => <FileModalScreenView files={[imageFile1, imageFile2]} initialIndex={1} />)
-  .add('video', () => <FileModalScreenView files={[videoFile]} initialIndex={0} />)
+  .add('image', () => <MainView files={[imageFile1, imageFile2]} initialIndex={1} />)
+  .add('video', () => <MainView files={[videoFile]} initialIndex={0} />)
